services/web/requirements.txt
Flask==1.1.1
psycopg2-binary==2.8.3
requests==2.22.0
